Preparing for major refactoring.
This is the "final version" -- v2.3.0 -- prior to a major refactoring.
This version is fully functional, though the "first run" setup is failing
and will be fixed as part of the refactoring.

Much of the refactoring is to prepare for migration to the Responses
API in addition to tightening up the code.